+++ pyslvs_ui/widgets/main_base.py
@@ -157,6 +157,7 @@ class Preferences:
     background_offset_x_option: float = 0
     background_offset_y_option: float = 0
     tick_mark_option: int = 1
+    nav_toolbar_pos_option: int = 1
     grab_no_background_option: bool = True
     monochrome_option: bool = False
     undo_limit_option: int = 60